+++ lib/pgsync/task.rb
@@ -88,7 +88,7 @@ module PgSync
 
       copy_to_command = "COPY (SELECT #{copy_fields} FROM #{quoted_table}#{sql_clause}) TO STDOUT"
       if opts[:in_batches]
-        raise Error, "Table must have a primary key to use --in-batches" if primary_key.empty?
+        raise Error, "Primary key required for --in-batches" if primary_key.empty?
         primary_key = primary_key.first
 
         destination.truncate(table) if opts[:truncate]
@@ -125,7 +125,11 @@ module PgSync
           end
         end
       elsif !opts[:truncate] && (opts[:overwrite] || opts[:preserve] || !sql_clause.empty?)
-        raise Error, "Table must have a primary key to use this functionality" if primary_key.empty?
+        if primary_key.empty?
+          raise Error, "Primary key required for --overwrite" if opts[:overwrite]
+          raise Error, "Primary key required for --preserve" if opts[:preserve]
+          raise Error, "Primary key required to sync specific rows"
+        end
 
         # create a temp table
         temp_table = "pgsync_#{rand(1_000_000_000)}"